+++ gulpfile.js
@@ -12,4 +12,5 @@ fs.readdirSync(tasksDir).forEach(filename => gulp.task(
 ));
 
 gulp.task('test', gulpsync.sync(['build-tests', 'test-runner']));
+gulp.task('e2e-test', gulpsync.sync(['e2e-test-runner']));
 gulp.task('default', done => runSequence('lint', 'test', 'clean', 'build', done));