+++ src/cli/commands/public-cmds/import-cmd.js
@@ -11,10 +11,8 @@ import { ComponentDependencies } from '../../../scope';
 export default class Import extends Command {
   name = 'import [ids...]';
   description = 'import a component';
-  alias = 'i';
+  alias = '';
   opts = [
-    ['s', 'save', 'save into bit.json'],
-    ['e', 'environment', 'install development environment dependencies (compiler | tester)'],
     ['t', 'tester', 'import a tester environment component'],
     ['v', 'verbose', 'show a more verbose output when possible'],
     ['c', 'compiler', 'import a compiler environment component'],
@@ -22,14 +20,12 @@ export default class Import extends Command {
   ];
   loader = true;
 
-  action([ids, ]: [string[], ], { save, tester, compiler, verbose, prefix, environment }:
+  action([ids, ]: [string[], ], { tester, compiler, verbose, prefix }:
   {
-    save?: bool,
     tester?: bool,
     compiler?: bool,
     verbose?: bool,
     prefix?: bool,
-    environment?: bool,
   }): Promise<any> {
     if (prefix) { return Promise.reject(new Error('prefix option currently not supported')); }
     // TODO - prefix returns true instead of the relevant string.
@@ -37,8 +33,11 @@ export default class Import extends Command {
     if (tester && compiler) {
       throw new Error('you cant use tester and compiler flags combined');
     }
+    if (!ids.length) {
+      console.log(chalk.yellow('\nwarning - using "bit import" without Ids is deprecated. Please use "bit install" instead\n')); // eslint-disable-line
+    }
 
-    return importAction({ ids, save, tester, compiler, verbose, prefix, environment });
+    return importAction({ ids, tester, compiler, verbose, prefix, environment: false });
   }
 
   report({ dependencies, envDependencies, warnings }: {