+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - add `bit reset` command for cancelling the last local commit
+- extract the importing bit.json components functionality from `bit import` into a new command `bit install`.
+- add infrastructure for e2e tests    
+- fix onExport hook to get called after writing dependencies to bit.json
 - increased max listeners to 100 (prevent warning message)
 - colored commit success message
 - support for merge conflict error reporting via ssh